requirements.txt
# list of all essential python dependencies for the project
django
djangorestframework
django-cors-headers
python-dotenv
djangorestframework-simplejwt